pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codefun-autosubmit"
version = "2.0.0"
authors = [
    {name = "KNN-07", email = "nktrungkien@protonmail.com"},
]
description = "Automated submission tool for Codefun.vn"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "pyperclip==1.11.0",
    "python-dotenv==1.2.1",
    "selenium==4.39.0",
    "requests==2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/KNN-07/codefun-autosubmit"
Documentation = "https://github.com/KNN-07/codefun-autosubmit#readme"
Repository = "https://github.com/KNN-07/codefun-autosubmit.git"
"Bug Tracker" = "https://github.com/KNN-07/codefun-autosubmit/issues"

[project.scripts]
codefun-autosubmit = "codefun_autosubmit.cli:main"
codefun = "codefun_autosubmit.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["codefun_autosubmit*"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=codefun_autosubmit --cov-report=html --cov-report=term-missing"